tamp

+++ database/migrations/2024_06_27_142529_formal.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->dateTime('tanggalmulai');
             $table->dateTime('tanggalselesai');
             $table->text('gelar');
-            $table->integer('IPK');
+            $table->double('IPK');
             $table->text('deskripsi');
             $table->timestamps();
         });
